% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chicago_arson}
\alias{chicago_arson}
\title{Chicago Arson}
\format{
A csv file with 522 coordinates and 3 variables:
\describe{
  \item{X}{X coordindate}
  \item{Y}{Y coordindate}
  \item{date}{date of event}
}
}
\usage{
chicago_arson
}
\description{
Arson events csv file. See vignette prepare_data for details.
}
\keyword{datasets}
