% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chicago_be}
\alias{chicago_be}
\title{Chicago Breaking and Entering}
\format{
A csv file with 11701 coordinates and 3 variables:
\describe{
  \item{X}{X coordindate}
  \item{Y}{Y coordindate}
  \item{date}{date of event}
}
}
\usage{
chicago_be
}
\description{
Residential burglary/breaking & entering events csv file. See vignette
prepare_data for details.
}
\keyword{datasets}
