% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_near_repeat.R
\name{tidy_near_repeat}
\alias{tidy_near_repeat}
\title{Calculate near-repeats using the Knox test with tidy inputs and outputs}
\usage{
tidy_near_repeat(
  data,
  x = NULL,
  y = NULL,
  time = NULL,
  sds = c(0, 200, 400, 600, 800, 1000),
  tds = c(0, 7, 14, 21, 28),
  ...
)
}
\arguments{
\item{data}{A data frame, tibble or Simple Features (SF) object in which each
row represent a single event.}

\item{x, y}{The names of the numeric columns in \code{data} that contain the
\code{x} (longitude, easting, etc.) and \code{y} (latitude, northing, etc.)
co-ordinates for each event. If \code{x} and \code{y} are not specified and
\code{data} is an SF object, the function will attempt to extract
co-ordinates from the geometry column.}

\item{time}{The bare name of the column in \code{data} containing the date
of each event. This column must contain Date values. If not given and the
data contains only one column of the type Date, the function will use this
column automatically.}

\item{sds}{A numeric vector of distances to be used as the breaks between
distance bands over which to test for near-repeat events. These distances
should be expressed in the same units (metres, feet, etc.) as used to
specify co-ordinates in the \code{x} and \code{y} columns.}

\item{tds}{A numeric vector of numbers of days to be used as the breaks
between temporal bands over which to test for near-repeat events.}

\item{...}{Optional arguments passed to \code{\link{NearRepeat}}}
}
\value{
A tibble containing observed counts and Knox test results for each
  combination of distance and time bands.
}
\description{
A wrapper around \code{\link{NearRepeat}} that accepts inputs and produces
outputs that meet the definition of
\href{https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html}{tidy data}
and work within a pipeline of functions.
}
\details{
This function uses default values for spatial and temporal distance bands
that are likely to be reasonable in many circumstances \strong{if
co-ordinates are given in metres}, as well as warning the user when the
values they have provided may produce unexpected results.
}
